+++ CHANGELOG.md
@@ -366,6 +366,7 @@
 * ViewModel now collides with ViewModel
 * Moved NavigationNodeEx out of experimental into World (as NavigationNode)
 * Fixed ResourceFolder ExploreForType ignoring redirects (fixes GetResources and a few other things)
+* Added GetHandlesAll API to resource management
 * Added conditional/microscript resolvers for PlayerFlags and SessionFlags (conditional only) [Experimental]
 * Added ScriptStringSubber that calls scripts and substitutes the result [Experimental]
 * DelayScaled can now be used from threads other than the main thread [Untested]